Refactor valcommit
