+++ protocol/src/asset.ts
@@ -1,0 +1,6 @@
+import { G } from "./curve";
+
+export function getV(asset: string) {
+  const V = G.multiply(BigInt(asset));
+  return V;
+}
